requirements.txt
asgiref==3.5.0
cffi==1.15.0
cryptography==36.0.2
Django==4.0.4
djangorestframework==3.13.1
pycparser==2.21
PyJWT==2.3.0
pytz==2022.1
sqlparse==0.4.2
